+++ docs/change_log.rst
@@ -42,6 +42,10 @@ Unreleased
 The following are only available within Stem's `git repository
 <download.html>`_.
 
+ * **Controller**
+
+  * Added `support for NETWORK_LIVENESS events <api/response.html#stem.response.events.NetworkLivenessEvent>`_ (:spec:`44aac63`)
+
 .. _version_1.4:
 
 Version 1.4 (May 13th, 2015)